+++ packages/connection/docs/benchmark.txt
@@ -1,33 +1,33 @@
 RPC BENCHMARK
 ┌─────────┬──────────────────────────────────────────────┬───────────┬────────────────────┬──────────┬─────────┐
-│ (index) │ Task Name                                    │ ops/sec   │ Average Time (ns)  │ Margin   │ Samples │
+│ (index) │                  Task Name                   │  ops/sec  │ Average Time (ns)  │  Margin  │ Samples │
 ├─────────┼──────────────────────────────────────────────┼───────────┼────────────────────┼──────────┼─────────┤
-│ 0       │ '1+2: [sumi rpc] over socket'                │ '61,533'  │ 16251.352393412293 │ '±0.46%' │ 123067  │
-│ 1       │ '1+2: [sumi rpc] over message port'          │ '87,842'  │ 11384.057876311219 │ '±0.60%' │ 175685  │
-│ 2       │ '1+2: [json rpc] over message port'          │ '102,526' │ 9753.584853672035  │ '±0.67%' │ 205053  │
-│ 3       │ '1+2: [sumi rpc] over ws channel'            │ '70,605'  │ 14163.138458484722 │ '±0.93%' │ 141212  │
-│ 4       │ '1+2: [json rpc] over ws channel'            │ '59,373'  │ 16842.59470134079  │ '±0.60%' │ 118747  │
-│ 5       │ 'string(584b): [sumi rpc] over socket'       │ '58,776'  │ 17013.707332011792 │ '±0.44%' │ 117553  │
-│ 6       │ 'string(584b): [sumi rpc] over message port' │ '73,524'  │ 13600.850227814266 │ '±1.11%' │ 147050  │
-│ 7       │ 'string(584b): [json rpc] over message port' │ '83,052'  │ 12040.622022219    │ '±0.57%' │ 166105  │
-│ 8       │ 'string(200k): [sumi rpc] over socket'       │ '4,725'   │ 211609.0040203138  │ '±1.04%' │ 9452    │
-│ 9       │ 'string(200k): [sumi rpc] over message port' │ '8,795'   │ 113689.32588675804 │ '±2.63%' │ 17592   │
-│ 10      │ 'string(200k): [json rpc] over message port' │ '1,651'   │ 605491.9125302713  │ '±0.64%' │ 3304    │
-│ 11      │ 'string(1m): [sumi rpc] over socket'         │ '773'     │ 1292818.612403102  │ '±5.43%' │ 1548    │
-│ 12      │ 'string(1m): [sumi rpc] over message port'   │ '1,890'   │ 528879.8521946594  │ '±2.52%' │ 3782    │
-│ 13      │ 'string(1m): [json rpc] over message port'   │ '341'     │ 2924756.46783628   │ '±0.31%' │ 684     │
-│ 14      │ 'buffer(1m): [sumi rpc] over message port'   │ '2,237'   │ 446831.42975206283 │ '±2.19%' │ 4477    │
-│ 15      │ 'buffer(1m): [sumi rpc] over socket'         │ '715'     │ 1396797.3093575428 │ '±5.38%' │ 1432    │
-│ 16      │ 'buffer(1m): [json rpc] over message port'   │ '18'      │ 52632514.28947344  │ '±1.47%' │ 38      │
+│    0    │        '1+2: [sumi rpc] over socket'         │ '74,506'  │ 13421.61067473848  │ '±0.42%' │ 149014  │
+│    1    │     '1+2: [sumi rpc] over message port'      │ '103,105' │ 9698.795492548768  │ '±1.01%' │ 206212  │
+│    2    │     '1+2: [json rpc] over message port'      │ '118,124' │  8465.67609839376  │ '±0.61%' │ 236249  │
+│    3    │      '1+2: [sumi rpc] over ws channel'       │ '93,725'  │ 10669.46241947974  │ '±0.44%' │ 187451  │
+│    4    │      '1+2: [json rpc] over ws channel'       │ '71,292'  │ 14026.769337816475 │ '±0.55%' │ 142585  │
+│    5    │    'string(584b): [sumi rpc] over socket'    │ '68,728'  │ 14550.089847394827 │ '±0.40%' │ 137457  │
+│    6    │ 'string(584b): [sumi rpc] over message port' │ '91,667'  │ 10909.029556217403 │ '±1.27%' │ 183335  │
+│    7    │ 'string(584b): [json rpc] over message port' │ '92,300'  │  10834.1886217631  │ '±0.59%' │ 184602  │
+│    8    │    'string(200k): [sumi rpc] over socket'    │  '4,996'  │ 200149.60007375036 │ '±1.31%' │  9993   │
+│    9    │ 'string(200k): [sumi rpc] over message port' │  '8,854'  │ 112935.08719751755 │ '±2.82%' │  17710  │
+│   10    │ 'string(200k): [json rpc] over message port' │  '1,743'  │ 573578.8863459859  │ '±0.62%' │  3488   │
+│   11    │     'string(1m): [sumi rpc] over socket'     │   '858'   │ 1165283.9910032866 │ '±4.79%' │  1717   │
+│   12    │  'string(1m): [sumi rpc] over message port'  │  '2,404'  │ 415815.28112447186 │ '±2.29%' │  4810   │
+│   13    │  'string(1m): [json rpc] over message port'  │   '348'   │ 2866380.2386013344 │ '±0.50%' │   698   │
+│   14    │  'buffer(1m): [sumi rpc] over message port'  │  '2,582'  │ 387240.8667469301  │ '±1.85%' │  5165   │
+│   15    │     'buffer(1m): [sumi rpc] over socket'     │   '821'   │ 1217579.0301489525 │ '±4.75%' │  1643   │
+│   16    │  'buffer(1m): [json rpc] over message port'  │   '21'    │ 47464365.34925949  │ '±0.57%' │   43    │
 └─────────┴──────────────────────────────────────────────┴───────────┴────────────────────┴──────────┴─────────┘
 ONE OF BENCHMARK
 ┌─────────┬────────────────────────────────────────┬─────────────┬────────────────────┬──────────┬─────────┐
-│ (index) │ Task Name                              │ ops/sec     │ Average Time (ns)  │ Margin   │ Samples │
+│ (index) │               Task Name                │   ops/sec   │ Average Time (ns)  │  Margin  │ Samples │
 ├─────────┼────────────────────────────────────────┼─────────────┼────────────────────┼──────────┼─────────┤
-│ 0       │ 'one of'                               │ '495,826'   │ 2016.836323794656  │ '±0.22%' │ 991653  │
-│ 1       │ 'one of json'                          │ '242,226'   │ 4128.365720996987  │ '±0.14%' │ 484454  │
-│ 2       │ 'one of serialize'                     │ '1,008,091' │ 991.9734652062606  │ '±0.23%' │ 2016183 │
-│ 3       │ 'one of json stringify'                │ '633,263'   │ 1579.1221474153353 │ '±0.20%' │ 1266527 │
-│ 4       │ 'one of serialize without buffer'      │ '1,236,440' │ 808.7735046687874  │ '±0.23%' │ 2472881 │
-│ 5       │ 'one of json stringify without buffer' │ '978,521'   │ 1021.950083365123  │ '±0.19%' │ 1957043 │
+│    0    │                'one of'                │  '520,981'  │ 1919.4552346553628 │ '±0.13%' │ 1041963 │
+│    1    │             'one of json'              │  '250,538'  │ 3991.4021558560885 │ '±0.08%' │ 501078  │
+│    2    │           'one of serialize'           │ '1,046,149' │ 955.8860568410437  │ '±0.12%' │ 2092300 │
+│    3    │        'one of json stringify'         │  '642,650'  │ 1556.0563739421286 │ '±0.13%' │ 1285301 │
+│    4    │   'one of serialize without buffer'    │ '1,285,842' │  777.699943253875  │ '±0.13%' │ 2571686 │
+│    5    │ 'one of json stringify without buffer' │ '1,002,895' │ 997.1133133039771  │ '±0.10%' │ 2005791 │
 └─────────┴────────────────────────────────────────┴─────────────┴────────────────────┴──────────┴─────────┘
