+++ packages/connection/src/common/capturer.ts
@@ -1,0 +1,36 @@
+declare global {
+  interface Window {
+    __OPENSUMI_DEVTOOLS_GLOBAL_HOOK__: any;
+  }
+}
+
+export enum MessageType {
+  SendNotification = 'sendNotification',
+  SendRequest = 'sendRequest',
+  RequestResult = 'requestResult',
+  OnNotification = 'onNotification',
+  OnRequest = 'onRequest',
+  OnRequestResult = 'onRequestResult',
+}
+
+export enum ResponseStatus {
+  Success = 'success',
+  Fail = 'fail',
+}
+
+export interface ICapturedMessage {
+  type: MessageType;
+  serviceMethod: string;
+  arguments?: any;
+  requestId?: string;
+  status?: ResponseStatus;
+  data?: any;
+  error?: any;
+}
+
+export function getCapturer() {
+  if (typeof window !== 'undefined' && window.__OPENSUMI_DEVTOOLS_GLOBAL_HOOK__?.captureRPC) {
+    return window.__OPENSUMI_DEVTOOLS_GLOBAL_HOOK__.captureRPC;
+  }
+  return;
+}
