+++ packages/connection/src/common/index.ts
@@ -2,7 +2,7 @@ export * from './message';
 export * from './rpc-service/proxy';
 export * from './rpc/multiplexer';
 export * from './rpcProtocol';
-export * from './utils';
+export * from './capturer';
 export * from './ws-channel';
 export * from './connect';
 export * from './types';
