+++ packages/connection/src/common/rpc/index.ts
@@ -1,2 +1,2 @@
 export * from './types';
-export * from './protocol-repository';
+export * from './message-io';
